Continued progress on streamlining FFT interface
Retain ability to work with actual complex Fourier coeffs, as this is a useful fallback/test option. Currently this is the only option that is actually working.

Moved this over to using complex FFTs throughout (the previous version did a weird hybrid where at the last moment, the offset fields were stored in real format, which makes debugging/comparisons much less clear).