WIP to allow for "stray particles", defined as high-res particles that are outside the high-res grid
This is partially complete. The key enabling class is ResolutionMatchingGrid which combines a
high-res grid with a supersampled low-res grid in one seamless view.

The mapping for such a situation now works, with a new command enabling the
mode. However, generating particles turns out to be very hard - it requires decoupling the grid
on which Zeldovich offsets were calculated from the grid on which the particles are generated.

At the current working point, this is partially completed by creating a set of classes to
evaluate fields (evaluator.hpp). Actually, these serve to decouple some of the functionality
in the Grid classes which uncomfortably knew too much about fields. Similar functionality
split has been applied to the Zeldovich particle generation classes. However, work still
needs to be done on the appropriate instantiation procedures, as they still have the
one-to-one grid-mapping baked in.