Complete work on stray particle system, which is now fully operational.
This update involves a significant rework of the way fields and particles are evaluated, breaking these systems into separate "evaluator" and "generator/storage" classes. The former take responsibility for evaluating on any specified grid; the latter only generate the underlying information and are no longer used for direct access to the results.