Correct reference output for test_02g_multilevelvariance
The unit test failure is related to test_02g_multilevelvariance which opens a zoom grid at coordinates (0,0,0), meaning that the coarse particle with ID 0 is actually 'zoomed'.

The TwoLevelParticleMapper class in the master branch does not account for this possibility. It tacitly assumes that particle 0 is part of the output. This means that the reference output for this test actually contains the wrong set of coarse particles.

The updated TwoLevelParticleMapper correctly checks for this possibility in begin() and so in fact is generating the correct set of coarse particles.