Fix mapper tests after changing behaviour.
The new behaviour flags all "zoomed" particles when a new level is opened. Previously, this would not happen.
In nearly all cases, this shouldn't matter, but there are edge cases where it does which were captured by the
tests. Additionally, the tests actually flagged a bug where levels referenced by the input mapper but not
the main mapper could cause incorrect behaviour.