Rename grid pointsToGrid -> isProxyFor
This is an attempt to be more specific that it's whether the grid "stands in for" rather than "knows about" the other grid.

That distinction becomes critical with the ResolutionMatchingGrid case, when it is being subsampled/supersampled but "knows about" multiple levels.
It's the level it's standing in for that we actually care about.