Implement automatic inter-level padding of sharp resolution contrasts
Specifically, when there is a sharp resolution contrast, make virtual grids and embed them
appropriately within the particle mapper to create a buffer zone of a user-specified number
of cells wide.

This requires some significant updates to core functionality, including the ability to
flag cells on a SubSampleGrid, SuperSampleGrid etc _independently_ of the underlying
grid. This functionality has been implemented using a new virtual grid. It has been
adopted as a wrapper around the ResolutionMatchingGrid in the case of 'stray particle'
mode, meaning particles can be mapped more accurately in simulations with strays.

It also leads to situations where one has subsample grids constructed on top of supersample
grids, so new logic added to get back a more efficient representation of the sampling
in such situations.