Continue clarifying code ready for release
This update focuses on removing some redundancy with the way covariances on grids are stored (delegating it
fully to the CAMB, i.e. input power spectrum interpolator, class). It also continues to move towards using
an enum rather than an int across the code to represent which type of transfer function is in use.