Fix importing overdensity fields
This had broken because we now keep the fields as whitenoise until later in the process
As a knock-on effect, it's impossible to retain an overall mean in the field (since
this should have zero covariance when multipled by the power spectrum). This
necessitates updating the test to a zero-mean random field.