Fix bug with double-zooms near the edge of a domain
Specifically, when calculating zoom windows, it was assumed that wrapping was available. But
when a second-level zoom is opened, it is embedded in a domain that cannot be wrapped.
The code would calculate a window that was only valid if wrapping were available, then
correctly complain that the window lay outside the grid above.

Solution is to clamp the window so that it can't go outside the grid above unless that
grid is in fact the base-level grid.

Add test